Initial test with half3, not working, complains about data not being of length 4

+++ Assets/Scripts/Vertex.cs
@@ -5,15 +5,15 @@ using System.Runtime.InteropServices;
 [StructLayout(LayoutKind.Sequential)]
 public struct Vertex
 {
-    public float3 pos;
-    public float3 normal;
+    public half3 pos;
+    public half3 normal;
 
     public static VertexAttributeDescriptor[] GetVertexAttributes()
     {
         return new[]
         {
-            new VertexAttributeDescriptor(VertexAttribute.Position, VertexAttributeFormat.Float32, 3),
-            new VertexAttributeDescriptor(VertexAttribute.Normal, VertexAttributeFormat.Float32, 3),
+            new VertexAttributeDescriptor(VertexAttribute.Position, VertexAttributeFormat.Float16, 3),
+            new VertexAttributeDescriptor(VertexAttribute.Normal, VertexAttributeFormat.Float16, 3),
         };
     }
 }
